debugged for 592 * 2048 - needs cleanup

+++ cudaFFT/cudaFFTwrapper.h
@@ -6,5 +6,5 @@ void ComputeTheFFT(std::complex<float>* h_signal, std::complex<float>* h_signal_
 void addjustCoefficientMagnitude(std::complex<float>* h_data, long dataSize) noexcept;
 int isOriginalEqualToTheTransformedAndInverseTransformenData(
     const std::complex<float>* original, const std::complex<float>* transformed, long dataSize) noexcept;
-void printTheData(const std::complex<float>* original, const std::complex<float>* transformed, long dataSize, const int printOffset = 0);
+void printTheData(const std::complex<float>* original, const std::complex<float>* transformed, long dataSize, const size_t printOffset = 0);
 void initializeTheSignals(std::complex<float>* fft, long dataSize) noexcept;
